+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/altsound/AltSoundServiceClient.java
@@ -3,6 +3,7 @@ package de.mephisto.vpin.restclient.altsound;
 import de.mephisto.vpin.restclient.assets.AssetType;
 import de.mephisto.vpin.restclient.client.VPinStudioClient;
 import de.mephisto.vpin.restclient.client.VPinStudioClientService;
+import de.mephisto.vpin.restclient.games.descriptors.UploadDescriptor;
 import de.mephisto.vpin.restclient.jobs.JobExecutionResult;
 import de.mephisto.vpin.restclient.util.FileUploadProgressListener;
 import org.slf4j.Logger;
@@ -47,10 +48,10 @@ public class AltSoundServiceClient extends VPinStudioClientService {
     return getRestClient().delete(API + "altsound/" + gameId);
   }
 
-  public JobExecutionResult uploadAltSound(File file, String uploadType, int gameId, FileUploadProgressListener listener) throws Exception {
+  public UploadDescriptor uploadAltSound(File file, int gameId, FileUploadProgressListener listener) throws Exception {
     try {
       String url = getRestClient().getBaseUrl() + API + "altsound/upload";
-      ResponseEntity<JobExecutionResult> exchange = new RestTemplate().exchange(url, HttpMethod.POST, createUpload(file, gameId, uploadType, AssetType.ALT_SOUND, listener), JobExecutionResult.class);
+      ResponseEntity<UploadDescriptor> exchange = new RestTemplate().exchange(url, HttpMethod.POST, createUpload(file, gameId, null, AssetType.ALT_SOUND, listener), UploadDescriptor.class);
       return exchange.getBody();
     } catch (Exception e) {
       LOG.error("ALT sound upload failed: " + e.getMessage(), e);