+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/games/descriptors/UploadDescriptorFactory.java
@@ -5,16 +5,20 @@ import org.springframework.web.multipart.MultipartFile;
 public class UploadDescriptorFactory {
 
   public static UploadDescriptor error(String message) {
-    UploadDescriptor descriptor  = new UploadDescriptor();
+    UploadDescriptor descriptor = new UploadDescriptor();
     descriptor.setError(message);
     return descriptor;
   }
 
   public static UploadDescriptor create(MultipartFile file, int gameId) {
-    UploadDescriptor descriptor  = new UploadDescriptor();
+    UploadDescriptor descriptor = new UploadDescriptor();
     descriptor.setGameId(gameId);
     descriptor.setFile(file);
     descriptor.setOriginalUploadedFileName(file.getOriginalFilename());
     return descriptor;
+  }
+
+  public static UploadDescriptor create(MultipartFile file) {
+    return create(file, 0);
   }
 }