+++ vpin-studio-server/src/main/java/de/mephisto/vpin/server/puppack/PupPacksResource.java
@@ -1,7 +1,10 @@
 package de.mephisto.vpin.server.puppack;
 
 import de.mephisto.vpin.connectors.vps.model.VpsDiffTypes;
+import de.mephisto.vpin.restclient.assets.AssetType;
 import de.mephisto.vpin.restclient.client.CommandOption;
+import de.mephisto.vpin.restclient.games.descriptors.UploadDescriptor;
+import de.mephisto.vpin.restclient.games.descriptors.UploadDescriptorFactory;
 import de.mephisto.vpin.restclient.jobs.JobExecutionResult;
 import de.mephisto.vpin.restclient.jobs.JobExecutionResultFactory;
 import de.mephisto.vpin.restclient.popper.PopperScreen;
@@ -9,6 +12,7 @@ import de.mephisto.vpin.restclient.puppacks.PupPackRepresentation;
 import de.mephisto.vpin.server.games.Game;
 import de.mephisto.vpin.server.games.GameService;
 import de.mephisto.vpin.server.games.GameValidationService;
+import de.mephisto.vpin.server.games.UniversalUploadService;
 import de.mephisto.vpin.server.util.UploadUtil;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
@@ -42,6 +46,9 @@ public class PupPacksResource {
 
   @Autowired
   private GameValidationService validationService;
+
+  @Autowired
+  private UniversalUploadService universalUploadService;
 
 
   @DeleteMapping("{id}")
@@ -103,30 +110,21 @@ public class PupPacksResource {
   }
 
   @PostMapping("/upload")
-  public JobExecutionResult upload(@RequestParam(value = "file", required = false) MultipartFile file,
-                                   @RequestParam(value = "uploadType", required = false) String uploadType,
-                                   @RequestParam("objectId") Integer gameId) {
+  public UploadDescriptor upload(@RequestParam(value = "file", required = false) MultipartFile file,
+                                 @RequestParam("objectId") Integer gameId) {
+    UploadDescriptor descriptor = UploadDescriptorFactory.create(file, gameId);
     try {
-      if (file == null) {
-        LOG.error("Upload request did not contain a file object.");
-        return JobExecutionResultFactory.error("Upload request did not contain a file object.");
-      }
-
-      Game game = gameService.getGame(gameId);
-      if (game == null) {
-        LOG.error("No game found for PUP pack upload.");
-        return JobExecutionResultFactory.error("No game found for PUP pack upload.");
-      }
-
-      String extension = FilenameUtils.getExtension(file.getOriginalFilename());
-      File pupTempArchive = File.createTempFile(FilenameUtils.getBaseName(file.getOriginalFilename()), "." + extension);
-      LOG.info("Uploading " + pupTempArchive.getAbsolutePath());
-      UploadUtil.upload(file, pupTempArchive);
-      JobExecutionResult jobExecutionResult = pupPacksService.installPupPack(game, pupTempArchive);
+      descriptor.getAssetsToImport().add(AssetType.PUP_PACK);
+      descriptor.upload();
+      universalUploadService.importArchiveBasedAssets(descriptor, AssetType.PUP_PACK);
       gameService.resetUpdate(gameId, VpsDiffTypes.pupPack);
-      return jobExecutionResult;
-    } catch (Exception e) {
-      throw new ResponseStatusException(INTERNAL_SERVER_ERROR, "PUP pack upload failed: " + e.getMessage());
+      return descriptor;
+    }
+    catch (Exception e) {
+      LOG.error(AssetType.PUP_PACK.name() + " upload failed: " + e.getMessage(), e);
+      throw new ResponseStatusException(INTERNAL_SERVER_ERROR, AssetType.PUP_PACK.name() + " upload failed: " + e.getMessage());
+    } finally {
+      descriptor.finalizeUpload();
     }
   }