+++ vpin-studio-ui/src/main/java/de/mephisto/vpin/ui/tables/dialogs/AltColorUploadProgressModel.java
@@ -1,6 +1,7 @@
 package de.mephisto.vpin.ui.tables.dialogs;
 
 import de.mephisto.vpin.commons.utils.WidgetFactory;
+import de.mephisto.vpin.restclient.games.descriptors.UploadDescriptor;
 import de.mephisto.vpin.restclient.jobs.JobExecutionResult;
 import de.mephisto.vpin.ui.Studio;
 import de.mephisto.vpin.ui.events.EventManager;
@@ -23,17 +24,13 @@ public class AltColorUploadProgressModel extends ProgressModel<File> {
   private final static Logger LOG = LoggerFactory.getLogger(AltColorUploadProgressModel.class);
 
   private final Iterator<File> iterator;
-  private final TablesSidebarController tablesSidebarController;
   private final int gameId;
   private final File file;
-  private final String altColorType;
 
-  public AltColorUploadProgressModel(TablesSidebarController tablesSidebarController, int gameId, String title, File file, String altColorType) {
+  public AltColorUploadProgressModel(int gameId, String title, File file, String altColorType) {
     super(title);
-    this.tablesSidebarController = tablesSidebarController;
     this.gameId = gameId;
     this.file = file;
-    this.altColorType = altColorType;
     this.iterator = Collections.singletonList(this.file).iterator();
   }
 
@@ -60,7 +57,7 @@ public class AltColorUploadProgressModel extends ProgressModel<File> {
   @Override
   public void processNext(ProgressResultModel progressResultModel, File next) {
     try {
-      JobExecutionResult result = Studio.client.getAltColorService().uploadAltColor(next, altColorType, gameId, percent -> progressResultModel.setProgress(percent));
+      UploadDescriptor result = Studio.client.getAltColorService().uploadAltColor(next, gameId, percent -> progressResultModel.setProgress(percent));
       if (!StringUtils.isEmpty(result.getError())) {
         Platform.runLater(() -> {
           WidgetFactory.showAlert(Studio.stage, "Error", result.getError());